Move test_avro.py
